Corman Algorithm benchmarked, Kahn improved
First of all the question arose, whether the input of the 'createRandomDAG'
script is to "easy". Easy in terms of the structure, which is could be
close to sorted input, if the algorithm traverses the nodes in ascend
order considering the node-id. To overcome this problem a new script has
been created, which creating "a bit more random" graphes, which containing
more "interesting" connections. However, it turned out, that this does not
lead to measurements, which would rank the algorithms in a different order.
However (second), the new script also contains "structure", even if those
one is more complex. A better way - which for some reasons has not been -
considered first, would be to create an empty matrix and add nodes which
does not violate the DAG property. Therefor:

1) Create a script, which "really creates random" DAGs.

Furthermore the implementations for the Kahn1962 algorithm has been
improved. Now this algorithm is in O(|V| + |E|).

2) The measurements revealed, Corman and Kahn seems to have some kind
of small log-factor. This should be investigated deeper.